blog/requirements.txt
asgiref==3.5.1
Django==4.0.4
djangorestframework==3.13.1
django-allauth==0.50.0
djangorestframework-simplejwt==5.1.0
dj-rest-auth==2.2.4
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
